Major rewrite, made it less "BASH" like
Rewrote a ton of stuff. Instead of manually editing the script to enter variables it now prompts the user for input. Create an xml config file where all that data is stored. Now depends on [PS-Pushover](https://github.com/Kieranties/PS-Pushover) for sending push notifications (only if configured to do so in script.)